Did alot
Created environment variables to store the cat api's endpoint and api keys. Added the domain to the next.js config file so the image would be authorized to load.
Created an api endpoint for geting. the cat breed data and getting cat images by search.

Made changed to the tailwind config to have variables for the font, colors and gradients
Adding tailwind styling while making use of the daisy ui premade class components.
Made a page to display the information and image of any particular cat breed.
Added funtionality to push the selected cat breed as a query in the search url, Making the cat breeds searchable byb directly editing the url.

Added a navigation bar to the layout react-typescript file so it would be shared by all pages.

Used axios in the home page to make a request to my endpoint for the cat breed data. Then extracted the cat breed names from each breed object from the data to make a select options list.