Working on implementing a breed rating system
Added validation for the url params.
Added a logical check to ensure the animal breed data was fetched successfully, If its not an error object is is passed in its place.

Added a default value to the select attribute.

Created a script to connect the app to  a mongo database

created a mongoose schema for the breed ratings to be stored in the right format in the database

Created and styled inputs by which a certain breed can be rated by the user

Made the breed search bar its own client component to redirect to the appropriate url when a breed is selected